requirements.txt
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
numpy
pandas
opencv-python
Pillow
logzero
boto3>=1.21.24
decord
typing
torch==1.11.0+cu113 
torchvision==0.12.0+cu113
scikit-image
transformers~=4.10.2
python-dotenv
ftfy
tqdm
regex
scenedetect
sentence-transformers
snowflake-connector-python
pyarrow==5.0.0
fastparquet